// packages/ui/src/components/types.ts
export interface IBaseButtonProps extends /* @vue-ignore */ ButtonHTMLAttributes {
	text?: string;
	theme?: "normal" | "info" | "plain" | "table" | "close" | "danger" | "navigation";
	icon?: object | string;
	loading?: boolean;
	disabled?: boolean;
}

export interface IBaseDialogProps {
	title?: string;
	subtitle?: string;
	bodyClass?: string;
	bodyPadding?: string;
	footerClass?: string;
	closable?: boolean;
	cancelConfig?: IBaseButtonProps;
}

export interface IBaseDialogEmits {
	(event: "close"): void;
	(event: "cancel"): void;
	(event: "open"): void;
}

export type ISortIdentity = -1 | 1;

export interface ITable<TData = unknown> extends ITanStackTable<TData> {
	getColumnSortIdentity: (columnId: string) => ISortIdentity;
	getSortedRowIndex: (columnId: string) => number;
}

export type ITableRow<TData = unknown> = ITanStackRow<TData>;

export type ITableCell<TData = unknown, TValue = unknown> = ITanStackCell<TData, TValue>;

export interface ITableCellContext<TData = unknown, TValue = unknown> extends Omit<ITanStackCellContext<TData, TValue>, "table"> {
	table: ITable<TData>;
}
// ...
export type ITableColumnOmit<TData = unknown, T = ITanStackColumn<TData>> = T extends ColumnDefBase<TData> ? Omit<T, "cell"> : never;

export type ITableColumn<TData = unknown> = ITableColumnOmit<TData> & {
	cell?: ColumnDefTemplate<ITableCellContext<TData>>;
};

export type ITableSort = ITanStackSort;

export type TLabelAlign = "left" | "top";

export type TTableExpandedState = ExpandedState;

export type TInputValue = string | number | null | undefined;

export interface IBaseTabProps {
	title: string;
	value?: string;
	disabled?: boolean;
	contentClasses?: string;
}

export interface IBaseTabsProps {
	tabs: IBaseTabProps[];
	orientation?: "horizontal" | "vertical";
}

export interface IChangeEvent<T extends HTMLElement> extends Event {
	target: T;
}

export interface IUseTableData<TData = unknown> {
	data: MaybeRef<TData[] | undefined>;
	columns: MaybeRef<ITableColumn<TData>[] | undefined>;
	sortInitial?: ITableSort;
	searchInitial?: string;
	paginated?: boolean;
	canExpand?: (row: ITableRow<TData>) => boolean;
}
// ...
export interface ITableAction<T> extends IBaseButtonProps {
	handler: (record: T) => void;
	canClick?: (record: T) => boolean;
}